Try factory boy/requirements.txt
Django==5.2.7
factory_boy==3.3.3
